pyproject.toml
[build-system]
requires = ["build", "setuptools-scm"]

[project]
dynamic = ["version"]
name = "message-ix-models"
description = "Tools for the MESSAGEix-GLOBIOM family of models"
authors = [{ name = "IIASA Energy, Climate, and Environment (ECE) Program" }]
maintainers = [
  { name = "Paul Natsuo Kishimoto", email = "mail@paul.kishimoto.name" },
  { name = "Fridolin Glatter", email = "glatter@iiasa.ac.at" },
]
readme = "README.rst"
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: Apache Software License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: R",
  "Topic :: Scientific/Engineering",
  "Topic :: Scientific/Engineering :: Information Analysis",
]
requires-python = ">=3.8"
dependencies = [
  "click",
  "colorama",
  # When the minimum is greater than the minimum via message_ix; e.g.
  # message_ix >= 3.4.0 → ixmp >= 3.4.0 → genno >= 1.6.0",
  "genno >= 1.18.1",
  "iam_units >= 2023.9.11",
  "message_ix >= 3.4.0",
  "pooch",
  "pyam-iamc >= 0.6",
  "pycountry",
  "PyYAML",
  "sdmx1 >= 2.8.0",
  "tqdm",
]

[project.urls]
homepage = "https://github.com/iiasa/message-ix-models"
repository = "https://github.com/iiasa/message-ix-models"
documentation = "https://docs.messageix.org/models"

[project.optional-dependencies]
docs = [
  # sphinx.ext.autosummary covers the test suite, so all test requirements
  "message_ix_models[tests]",
  "sphinx >= 4",
  "sphinx_rtd_theme",
  "sphinxcontrib-bibtex",
]
tests = [
  # For nbclient, thus nbformat
  "ixmp[tests]",
  "pytest",
  "pytest-cov",
  "pytest-xdist",
]

[project.scripts]
mix-models = "message_ix_models.cli:main"

[tool.coverage.report]
exclude_also = [
  # Don't complain about abstract methods, they aren't run
  "@(abc\\.)?abstractmethod",
  # Imports only used by type checkers
  "if TYPE_CHECKING:",
]

[tool.mypy]
exclude = ["doc/"]

[[tool.mypy.overrides]]
# Packages/modules for which no type hints are available
module = [
  "colorama",
  "message_data.*",
  "pooch",
  "pycountry",
  # Indirectly via message_ix
  # This should be a subset of the list in message_ix's pyproject.toml
  "matplotlib.*",
  "pandas.*",
  "pyam",
  # Indirectly via ixmp
  # This should be a subset of the list in ixmp's pyproject.toml
  "dask.*",
  "jpype",
  "memory_profiler",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
# Work around iiasa/ixmp#465 and similar
module = ["ixmp.*"]
no_implicit_optional = false

[tool.pytest.ini_options]
# Disable faulthandler plugin on Windows to prevent spurious console noise
addopts = "-p no:faulthandler --cov=message_ix_models --cov-report="
filterwarnings = "ignore:distutils Version classes.*:DeprecationWarning"

[tool.ruff]
select = ["C9", "E", "F", "I", "W"]

[tool.ruff.mccabe]
max-complexity = 13

[tool.setuptools.packages]
find = {}

[tool.setuptools_scm]
